Simplifiy best avalaibale date. Extract TransactionView
